Add a Bool flag to TDef for references
TDefs are now indexed by the number of free varibales as well as
whether they have references or not.

`TDef' 3 True` means there are three free variables and contains
unbounded references.

`TDef 0 False` means there are no free variables and all references
are resolved.

This allows us to recover totality when computing the Idris Type
of a TDef using the `Ty` function. Also should allow us to remove
the effectful compiling of a TDef in the haskell backend by first
resolving all references into a debrujin index and extending the
previous context with the new indices used by references.

This comes at the cost of the existing dependent parser which is
now basically broken until we find a way to fix it:

- Either only parse TDefR which makes references unusable
- Remove the dependent parse and write an index-free parser #175